:art: split code in 3 modules
dispatch code from correplaionPlus module in

mapAnalysis
diffMap
centralityAnalysis

modules
